pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "survival-analysis"
version = "1.0.0"
description = "Clinical trials duration survival analysis (Cox PH + Kaplan-Meier)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "lifelines>=0.28",
    "pandas>=2.0",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "scikit-learn>=1.4",
    "statsmodels>=0.14",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "patsy>=0.5",
    "scipy>=1.12",
]

[project.scripts]
sa-cox       = "survival_analysis.cli.cox:main"
sa-km        = "survival_analysis.cli.km:main"
sa-agents    = "survival_analysis.cli.agents:main"
sa-hr-sens   = "survival_analysis.cli.hr_sensitivity:main"
sa-clusters  = "survival_analysis.cli.clusters:main"
sa-quantiles = "survival_analysis.cli.quantiles:main"
sa-all       = "survival_analysis.cli.all:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
